pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schmekla"
version = "0.1.0"
description = "Structural Modeling Application with AI-Assisted Design"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Your Company"}
]
keywords = ["structural", "modeling", "BIM", "IFC", "Tekla", "CAD"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Developers",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "ifcopenshell>=0.7.0",
    "cadquery>=2.4.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "PySide6>=6.6.0",
    "pyvista>=0.43.0",
    "pyvistaqt>=0.11.0",
    "vtk>=9.3.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
    "typing_extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[project.scripts]
schmekla = "src.main:main"

[project.gui-scripts]
schmekla-gui = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
